+++ services/frontend/src/app/core/service/business/administration.service.ts
@@ -14,6 +14,7 @@ import {
     OrganizationUpsertRest,
     Subscriber,
     SubscriberCriteriaRest,
+    workspace,
 } from "../../interfaces/administration.interfaces";
 import { AdministrationDataService } from "../data/administration-data-service";
 
@@ -86,7 +87,15 @@ export class AdministrationService {
         return this.administrationDataService.postOrganization(body);
     }
 
-    getDomainSubscribers(body: { email: string }): Observable<DomainSubscribers[]> {
+    getDomainSubscribers(body: { email?: string }): Observable<DomainSubscribers[]> {
         return this.administrationDataService.getDomainSubscribers(body);
+    }
+
+    postUserWorkspace(body: {
+        subscriberId?: number;
+        name?: string;
+        status?: string;
+    }): Observable<workspace> {
+        return this.administrationDataService.postUserWorkspace(body);
     }
 }